explore doing a large amounts of asynchronous work like 1,000 and trying to do another separate work like getting nthPrime of a large number
notice that few threads (about 10) were created for the 1,000 tasks and they contended for resources blocking the thread and not allowing the nhtPrime to perform.

it's advisable to always use non blocking asynchronous APIs in such situations like Task.sleep() or the ones provided in apple's API such as URLSession, LocationManager etc.